moved the util functionality from users to array
